Implement edit by removing tag
